refactor(settings): 設定・オンボーディング画面のカード分割と共有コンポーネント化
- General 設定を Profile / Display / Language / DataManagement / About に分割
- ProfileFormFields, LanguageSelectField をオンボーディングと設定で共有
- SectionSaveStatus でセクション単位の保存状態表示
- SettingsOverview をナビゲーション専用に整理、returnTo 対応
- Onboarding のアバター blob URL 解放でメモリリーク対策
- セルフレビュー doc 追加 (docs/reviews/review-develop-20260315-3.md)

Made-with: Cursor
