+++ src/components/settings/AISettingsForm.tsx
@@ -17,6 +17,7 @@ import {
 import { useAISettingsForm } from "./useAISettingsForm";
 import { AISettingsFormServerSection } from "./AISettingsFormServerSection";
 import { AISettingsFormUserKeySection } from "./AISettingsFormUserKeySection";
+import { SectionSaveStatus } from "./SectionSaveStatus";
 import { getProviderById } from "@/types/ai";
 import { useTranslation } from "react-i18next";
 
@@ -25,8 +26,20 @@ interface AISettingsFormProps {
   embedded?: boolean;
 }
 
-export const AISettingsForm: React.FC<AISettingsFormProps> = ({ embedded = false }) => {
+/**
+ *
+ */
+export /**
+ *
+ */
+const AISettingsForm: React.FC<AISettingsFormProps> = ({ embedded = false }) => {
+  /**
+   *
+   */
   const { t } = useTranslation();
+  /**
+   *
+   */
   const {
     settings,
     availableModels,
@@ -34,6 +47,7 @@ export const AISettingsForm: React.FC<AISettingsFormProps> = ({ embedded = false
     isSaving,
     isTesting,
     testResult,
+    savedAt,
     showApiKey,
     setShowApiKey,
     useOwnKey,
@@ -49,8 +63,19 @@ export const AISettingsForm: React.FC<AISettingsFormProps> = ({ embedded = false
     handleReset,
   } = useAISettingsForm();
 
+  /**
+   *
+   */
   const currentProvider = getProviderById(settings.provider);
+  /**
+   *
+   */
   const currentModelId = settings.modelId || `${settings.provider}:${settings.model}`;
+
+  /**
+   *
+   */
+  const saveStatus = isSaving ? "saving" : savedAt != null ? "saved" : "idle";
 
   if (isLoading) {
     return (
@@ -72,6 +97,7 @@ export const AISettingsForm: React.FC<AISettingsFormProps> = ({ embedded = false
       )}
 
       <CardContent className={embedded ? "space-y-6 pt-0" : "space-y-6"}>
+        {embedded && saveStatus !== "idle" && <SectionSaveStatus status={saveStatus} />}
         {isServerMode && (
           <AISettingsFormServerSection
             serverModels={serverModels}
