+++ src/components/settings/SettingsOverview.tsx
@@ -11,12 +11,16 @@ const SECTIONS: { id: SectionId; icon: React.ReactNode }[] = [
   { id: "storage", icon: <ImageIcon className="h-5 w-5" /> },
 ];
 
+/**
+ *
+ */
 export interface SettingsOverviewProps {
-  summaries: Record<SectionId, string>;
+  /** No longer displayed; kept for backward compatibility. Overview is navigation-only. */
+  summaries?: Record<SectionId, string>;
 }
 
-/** Renders overview cards using precomputed summaries (avoids duplicate useSettingsSummaries subscriptions when used from Settings hub). */
-export const SettingsOverview: React.FC<SettingsOverviewProps> = ({ summaries }) => {
+/** Renders overview cards (navigation only). Section summaries are shown in SettingsSection. */
+export const SettingsOverview: React.FC<SettingsOverviewProps> = () => {
   const { t } = useTranslation();
   const [searchParams] = useSearchParams();
   const returnTo = searchParams.get("returnTo");
@@ -53,9 +57,6 @@ export const SettingsOverview: React.FC<SettingsOverviewProps> = ({ summaries })
               <div className="min-w-0 flex-1">
                 <CardTitle className="text-base">{getTitle(id)}</CardTitle>
                 <CardDescription className="text-sm">{getDescription(id)}</CardDescription>
-                {summaries[id] && (
-                  <p className="mt-1.5 text-xs font-medium text-foreground">{summaries[id]}</p>
-                )}
               </div>
             </CardHeader>
           </Card>
