+++ src/components/settings/useAISettingsForm.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback } from "react";
+import { useState, useEffect, useCallback, useRef } from "react";
 import { useTranslation } from "react-i18next";
 import { toast as sonnerToast } from "@zedi/ui/components/sonner";
 import { useToast } from "@zedi/ui";
@@ -8,8 +8,19 @@ import { fetchServerModels, FetchServerModelsError } from "@/lib/aiService";
 import type { AISettings } from "@/types/ai";
 import type { AIModel } from "@/types/ai";
 
+const SAVED_INDICATOR_MS = 3000;
+
+/**
+ *
+ */
 export function useAISettingsForm() {
+  /**
+   *
+   */
   const { t } = useTranslation();
+  /**
+   *
+   */
   const {
     settings,
     availableModels,
@@ -23,26 +34,64 @@ export function useAISettingsForm() {
     reset,
   } = useAISettings();
 
+  /**
+   *
+   */
   const [showApiKey, setShowApiKey] = useState(false);
+  /**
+   *
+   */
   const [useOwnKey, setUseOwnKey] = useState(false);
+  /**
+   *
+   */
   const [serverModels, setServerModels] = useState<AIModel[]>([]);
+  /**
+   *
+   */
   const [serverModelsLoading, setServerModelsLoading] = useState(false);
+  /**
+   *
+   */
   const [serverModelsError, setServerModelsError] = useState<string | null>(null);
+  /**
+   *
+   */
+  const [savedAt, setSavedAt] = useState<number | null>(null);
+  /**
+   *
+   */
+  const savedAtTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
+  /**
+   *
+   */
   const { toast } = useToast();
 
+  /**
+   *
+   */
   const isServerMode = settings.apiMode === "api_server" && !useOwnKey;
 
+  /**
+   *
+   */
   const loadServerModels = useCallback(
     async (forceRefresh = false) => {
       setServerModelsError(null);
       setServerModelsLoading(true);
       try {
+        /**
+         *
+         */
         const { models } = await fetchServerModels(forceRefresh);
         setServerModels(models ?? []);
         if (!models?.length) {
           setServerModelsError(t("aiSettings.modelsEmpty"));
         }
       } catch (e) {
+        /**
+         *
+         */
         const message =
           e instanceof FetchServerModelsError
             ? e.message
@@ -70,15 +119,21 @@ export function useAISettingsForm() {
     }
   }, [isLoading, settings.apiMode]);
 
+  /**
+   *
+   */
   const runSave = useCallback(async () => {
+    /**
+     *
+     */
     const success = await save();
     if (success) {
-      sonnerToast.success(t("aiSettings.savedToast"), {
-        description: t("aiSettings.savedToastDescription"),
-      });
-      // Do not navigate(returnTo) here: on the settings hub both AI and storage
-      // hooks are mounted, so auto-navigate would redirect after first save and
-      // cross-contaminate returnTo between sections. Use the hub's back button instead.
+      setSavedAt(Date.now());
+      if (savedAtTimeoutRef.current) clearTimeout(savedAtTimeoutRef.current);
+      savedAtTimeoutRef.current = setTimeout(() => {
+        setSavedAt(null);
+        savedAtTimeoutRef.current = null;
+      }, SAVED_INDICATOR_MS);
     } else {
       sonnerToast.error(t("common.error"), {
         description: t("aiSettings.saveFailedToastDescription"),
@@ -86,10 +141,19 @@ export function useAISettingsForm() {
     }
   }, [save, t]);
 
+  /**
+   *
+   */
   const scheduleSave = useDebouncedCallback(runSave, 800);
 
+  /**
+   *
+   */
   const updateSettings = useCallback(
     (updates: Partial<AISettings>) => {
+      /**
+       *
+       */
       const normalizedUpdates =
         (updates.provider !== undefined || updates.model !== undefined) &&
         updates.modelId === undefined
@@ -101,6 +165,9 @@ export function useAISettingsForm() {
     [updateSettingsBase, scheduleSave],
   );
 
+  /**
+   *
+   */
   const handleToggleOwnKey = useCallback(
     (checked: boolean) => {
       setUseOwnKey(checked);
@@ -111,7 +178,13 @@ export function useAISettingsForm() {
     [updateSettings],
   );
 
+  /**
+   *
+   */
   const handleTest = useCallback(async () => {
+    /**
+     *
+     */
     const result = await test();
     if (result.success) {
       toast({ title: t("aiSettings.connectionSuccess"), description: result.message });
@@ -124,6 +197,9 @@ export function useAISettingsForm() {
     }
   }, [test, toast, t]);
 
+  /**
+   *
+   */
   const handleReset = useCallback(() => {
     reset();
     setUseOwnKey(false);
@@ -133,8 +209,14 @@ export function useAISettingsForm() {
     });
   }, [reset, toast, t]);
 
+  /**
+   *
+   */
   const handleServerModelSelect = useCallback(
     (modelId: string) => {
+      /**
+       *
+       */
       const model = serverModels.find((m) => m.id === modelId);
       if (model) {
         updateSettings({
@@ -154,6 +236,7 @@ export function useAISettingsForm() {
     isSaving,
     isTesting,
     testResult,
+    savedAt,
     showApiKey,
     setShowApiKey,
     useOwnKey,
