+++ src/i18n/locales/en/settings.json
@@ -1,6 +1,8 @@
 {
   "title": "Settings",
   "hubDescription": "View and edit app settings such as theme, AI, and storage in one place.",
+  "saved": "Saved",
+  "saving": "Saving…",
   "general": {
     "title": "General",
     "description": "Theme, font size, language, profile"
