+++ src/i18n/locales/ja/settings.json
@@ -1,6 +1,8 @@
 {
   "title": "設定",
   "hubDescription": "テーマ、AI、ストレージなどアプリの設定を一覧で確認・編集できます。",
+  "saved": "保存しました",
+  "saving": "保存中…",
   "general": {
     "title": "一般設定",
     "description": "テーマ、フォントサイズ、言語、プロフィール"
